app/frontend/requirements.txt
streamlit>=1.31.0
requests
aiohttp
pydantic
nanoid
pillow
python-io
fastapi